Adding the last version of Nebula.
For the computer graphics' final project, my team and I developed Nebula. Nebula is intended to be a web app where the user can learn interesting astronomical data about the solar system and the universe in general, in an entertaining and attractive way.

Nebula was developed with HTML, CSS and JavaScript, using semantic framwork for creating an intuitive environment. The computer graphics related components were developed with ThreeJS.